[Core] Separating ResultDatabase from from_json_check_result_process
In order to use it in other classes without including from_json_check_result_process.h
